+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Experiment module: Duplicate proposals. [#659](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/659)
+
 ## [2.22.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/43)
 
 ### Changed